pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eight-layer-pqc"
version = "2.0.0"
description = "Enterprise-grade Eight-Layer Post-Quantum Cryptography Security Framework"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "TradeMomentum LLC", email = "security@trademomentum.com"}
]
maintainers = [
    {name = "TradeMomentum LLC", email = "security@trademomentum.com"}
]
keywords = [
    "post-quantum",
    "cryptography",
    "pqc",
    "security",
    "ml-kem",
    "ml-dsa",
    "fips-203",
    "fips-204",
    "quantum-resistant"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
requires-python = ">=3.10"
dependencies = [
    "pycryptodome>=3.19.0",
    "kyber-py>=1.0.0",
    "dilithium-py>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
]
docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/anumethod/eight-layer-pqc"
Documentation = "https://github.com/anumethod/eight-layer-pqc#readme"
Repository = "https://github.com/anumethod/eight-layer-pqc.git"
Issues = "https://github.com/anumethod/eight-layer-pqc/issues"

[tool.setuptools.packages.find]
where = ["code"]
include = ["python*"]

[tool.setuptools.package-dir]
"" = "code"
"eight_layer_pqc" = "code/python"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
